requirements/base.txt
# Non-django requirements
Fabric==1.8.1
numpy==1.8.0
IMAPClient==0.10.2
MySQL-python==1.2.5
Pillow>=2.5.1
docutils==0.9.1
geojson==1.0.1
html5lib==0.90
ipython>=2.0.0
matplotlib==1.0.1
pandas==0.14.1
Pint==0.5
py==1.4.9
pygeocoder==1.2.5
pyPdf==1.13
pytest==2.2.4
reportlab==2.7
requests==2.2.1
simplejson==3.3.2
simplekml==1.2.3
xhtml2pdf>=0.0.6
virtualenv==1.11.2
wsgiref==0.1.2

# Django requirements
Django==1.7.3
git+https://github.com/noirbizarre/django.js.git@65b267b04
django-ajax-selects==1.3.5
git+git://github.com/brack3t/django-braces.git@master
django-classy-tags==0.5.2
django_compressor>=1.4
git+git://github.com/girasquid/django-feedback.git
django-easy-pdf==0.1.0
django-floppyforms==1.3.0
django-imagekit==3.2.5
django-registration-redux==1.1
git+git://github.com/ebrelsford/django-tablib.git@master
django-widget-tweaks==1.3
django-xhtml2pdf==0.0.3
git+https://github.com/philippWassibauer/templated-emails.git@9f011ba84b
webplotlib==0.1
